Optimise resolution of fully bound concludables (#6243)
## What is the goal of this PR?
In some cases, a concludable that is being resolved may have all of its retrievable variables (named and anonymous) fully bound. This means that only a single answer is required to prove the concludable holds, rather than all possible ways. We add short circuiting to prevent exploring further rules in a concludable when a single answer has been found.

## What are the changes implemented in this PR?
* Concludables `RequestState` can record whether a single answer is required
* When producing an answer from a `ConcludableResolver`, we may short circuit and remove all possible downstream answer producers if only 1 answer is required

## ADR: why only short circuit in fully bound concludables
We could short circuit when any of our resolvers end up in a fully bound state. However, it's only useful in concludables:
* A fully bound `Retrievable` will only find exactly 1 answer, so short circuiting would not speed anything up
* `Negation` resolvers only ever look for 1 answer
* all `Conjunction` resolvers will very rarely be fully bound. Even if they are, if retrievables and concludables only return 1 answer when they are fully bound, then the conjunction too will only return 1 answer - so it doesn't add anything here

We could implement this optimisation for disjunctions, however, it is not worth our time at this moment as a fully bound disjunction is uncommon.