RPC Service Architecture Refactor (#6256)
## What is the goal of this PR?

Over time, our Client architecture, such as Client Java, has been extended considerably to implement various features and optimisations. It has lead the architecture to a complex codebase which is not so easy to follow and replicate in other languages. We have now refactored Client Java considerably to come up with an architecture that is easy to understand, easy to extend, and easy to replicate in other languages that wish to implement a Grakn Client. These changes on the client requires changes in the protocol that we need to apply onto the server-side RPC Service architecture.