Introduce ByteMap to manage indexes in Encoding class (#6264)
## What is the goal of this PR?

Recently we had introduced an optimisation to the Encoding class, where each encoding object can be efficiently retrieved by allowing lookup from an array indexed by the byte value of the encoding. This was a great idea, but the implementation was not foolproof. The logic of constructing the array and retrieving from the array, which were tied together were not scoped and abstracted within one class. This made the the usage of the array requires logic that is scattered throughout the Encoding class and error prone. In fact, an error did get introduced in the initial commit. By introducing `ByteMap` inner class in `Encoding` we are now able to concentrate the domain logic into one class and guarantee the correct behaviour.