Concludable and Retrievable Subsumption Caching (#6320)
## What is the goal of this PR?

We introduce caching in the reasoner avoid repeated work. Adding this caching has also allowed us to tighten the criteria upon which we trigger an outer reiteration of the reasoner for a given query. Such an iteration is necessary to search for inferred facts that can only be found based on the inferences made by prior iterations.

The cache ensures that work can be re-used for the same exact partial answer when revisited. We also implement cache subsumption: any partial answer can utilise the cached answers for any more general (subsuming) partial answer which has a complete cache.

## What are the changes implemented in this PR?

- Subsumptive caching for ConcludableResolver
- Subsumptive caching for RetrievableResolver
- Abstracts RequestState across the non-compound resolvers. The compound resolvers need to be brought in line with this approach in a follow-up PR.
- Adds Poller classes to implement polling the cache for new answers
- Account for the case of `$x isa $type` in rules (previously handled by always triggering at least one reiteration) by generating answers over the cartesian product of all possible types for the answer found. This takes place during un-unification.

Authored jointly with @flyingsilverfin