Request reiteration status with explicit resolver messages (#6332)
## What is the goal of this PR?

Within the message-passing framework of the reasoner, answers would carry a flag to indicate whether an additional outer iteration is necessary to find all answers. It was possible that the answers needed to carry this flag could be not present or swallowed, and so the answer produced would not reiterate. In recursive cases in particular this caused missed answers. We solve this by introducing request and response messages specifically to communicate whether reiteration is required.

## What are the changes implemented in this PR?

- Upon receiving a failed response, the ReasonerProducer asks all ConcludableResolvbles whether reiteration is required for the given RootResolver. It waits to receive a response from all of them before proceeding with reiteration or terminating.
- Adds ReiterationQuery Request and Response types
- Strips out `requiresReiteration` flags from `AnswerState`, since these are now redundant.