Deploy artifacts that include Console (#6352)
## What is the goal of this PR?

In order to test repositories that load TypeQL files into TypeDB, we require an artifact of TypeDB that includes the Console. This is because Console comes with built-in functionality for loading from TypeQL files, which cannot be easily replicated since it makes use of a TypeQL parser.

## What are the changes implemented in this PR?

Deploy artifacts for `typedb-all` for snapshots and releases.