Type casting errors include type label (#6347)
## What is the goal of this PR?

As identified in #6248 and in several community messages, we clean up UX that is exposed when a user mis-uses a type as a different meta type (eg. uses an entity in place of an attribute). We now include the type label in the error message to allow tracking down the error much more easily.

## What are the changes implemented in this PR?

* Type casting errors include the type label of the type that failed to cast