Fix rocks iterator is valid assertion error (#6375)
## What is the goal of this PR?

When RocksDB -dev is enabled (ie with its own assertions), tests (notable `BasicTest`) fail when re-opening a RocksDB database. This is caused by syncing our key generator, which relies on a rocks iterator's `seekForPrev()` call. We get an invalid iterator back when there are 0 keys in RocksDB, which is only a RocksDB error when development JNI is in use. We solve this by adding an `isValid()` check after doing `seekForPrev()`


## What are the changes implemented in this PR?

* if the iterator is invalid when calling `seekForPrev()`, we return null