Streamline statistics counting and fix ACID with no RocksDB writes (#6373)
## What is the goal of this PR?

As of #6306, Attribute PUT operations do a `read` before doing a `write` to rocksDB. We can harness this to only add a attribute PUT count job if the attribute is actually new, dramatically reducing the number of background count operations the statistics thread has to perform. 

This leads to an ACID bug where doing true attribute PUT leads to no actual writes on RocksDB, and therefore no storage version range against which to check for concurrent commits. We fix this by always writing at least one key to RocksDB per write transaction.


## What are the changes implemented in this PR?

* use the `isPersisted()` flag on an attribute vertex to only add a count job when putting an attribute that is actually brand new
* Always write a dummy key when committing a storage transaction. This guarantees that the RocksDB sequence number will increment and lead to a non-zero range that we can use for doing key-conflict detection and guaranteeing ACID safety.

Example of the problematic ACID case when statistics is writing fewer keys and no longer hiding this bug:
1. `name "Alice"` already exists in the storage, which is at sequence number 10
2. open write tx, put `name "Alice"`, commit. After commit, we are still at storage sequence number 10
3. open a write tx, delete `name "Alice"`. commit. This leads to a a conflict, because we opened a transaction at version 10 and a storage committed after or equal to version 10. However, it is not expected to fail due to using serialised transactions!

This becomes a non-issue if the sequence number always increments by at least one, because the second transaction opens at version 11. Then we can use the logic of checking only against storages that are committed greater than a version 10 (rather than equal to as well)

* To implement the dummy write we create a new `Encoding` prefix `System` and namespace `System` for future system-level operations
* Reserve a range of prefixes for use externally (eg. cluster).