Allow tests to include anonymous identifiers in filter (#6378)
## What is the goal of this PR?

For the purpose of testing reasoner, we need to be able to return answers to anonymous variables, for queries constructed using backend TypeQL patterns. This is almost doable, except that our filters only allow named identifiers. This change allows us to pass anonymous identifiers in the filter, and calling the `executeTraversal`/`executeReasoner` methods directly, which are now public.

## What are the changes implemented in this PR?

* expose to methods to call traversal/reasoning directly on `Reasoner` class
* allow passing anonymous identifiers in the query filter, which allows tests to receive anonymous variables' answers. This avoids the need to rewrite or deanonimize variables.