Reasoner Correctness Verification (#6374)
## What is the goal of this PR?

We introduce an internal framework for verifying the soundness and completeness of reasoning by comparing with a naive forward-chaining reasoner.

## What are the changes implemented in this PR?

### Initialisation

Use a simplistic forward-chaining method to create a reference transaction with all inferable facts materialised. While materialising record all of the facts that could be inferred all of the ways they could have been materialised.

### Soundness Verification

Check that all of the lines of reasoning taken are sound.

Run a query in a live reasoning transaction. For each answer explore the explanations depth-first. (A) Look up in the materialisation records of the materialised Tx the combination of when bindings plus the rule used to see that there's exactly one `then` answer. If there is no such lineage then the explanation is unsound and a rule was triggered incorrectly. If there is more than one materialisation then assert something in the framework is wrong.

Take the inferred `then` variable binding from the materialisation information. The inferred concepts in the two transactions can be mapped such that when verifying the next explanation in the tree the `when` is fully bound. Repeat from (A).

### Completeness Verification

From the query under test, split into concludables and find the relevant materialisations for each concludable. Do this in a depth first manner through the rules also. This finds the lines of reasoning needed to arrive at the specific query's answers. 

Starting from the leaf materialisations, check that the rule conclusion that should be reached has a single matching answer when queried in the live reasoning tx (since a rule does a put, and the `when` is fully bound there must be one answer). Next check that the concludable used to reach that rule has one or more answers in the live reasoning tx (since a concludable could be answered by a more specific pattern there can be more than one answer).

This approach has a limitation: we cannot verify concludables or conclusions that contain more than one inferred variable. This includes the cases that an inferred relation owning an inferred attribute, or an inferred relation relates an inferred concept. This is because we cannot also verify the additional inferred concepts without an onerous computational expense.

### BDD Resolution Steps

We also notice that the BDD scenarios all follow a distinct pattern for resolution tests, so we overhaul the steps to simplify them.

### Related Issues and PRs

Solves #6362 and #6384.
Changes to tests are made in https://github.com/vaticle/typedb-behaviour/pull/206

Failing verification steps are present, there's an issue for that here: https://github.com/vaticle/typedb/issues/6396