Create clear structure of resolver types (#6410)
## What is the goal of this PR?

We see that the resolvers share common patterns which can be abstracted. We also see that within the Concludable resolver there are distinct code paths depending upon whether the user's query is a match or explain query. It's now possible, based on prior refactoring, to split match and explain code paths into two separate classes.

## What are the changes implemented in this PR?

- Create a `Coordinator` class which encapsulates the main operations of `ConcludableResolver` and `RetrievableResolver`.
- Create `MatchBoundConcludableResolver` and `ExplainBoundConcludableResolver` since they have separate behaviours. This means explaning and matching now no longer share a cache.

Part of #6356