Cycle blocking and cycle re-exploration (#6429)
Cycles within reasoning need to be terminated to prevent infinite computation. These cycles need to be revisited when new answers are found at the start of the cycle. We introduce a new actor-centric mechanism for detecting these cycles and determining when they require revisiting. This replaces the preexisting method, which was to reiterate all reasoning actions to reveal new facts from the cycles.

- Exact cycle detection.
- A new `blocked` response type to indicate that a cycle was encountered.
- A new `revisit` request type to trigger downstream unblocking.
- Downstream paths can be set to blocked by a given set of cycles, unblocked when new answers are encountered or a relevant revisit request is received.
- A `Template` Request from which `Visit` or `Revisit` requests can be built with a supplied `Trace`.
- Remove much of the subsumption code since this wasn't in a working state.
- Add a replay buffer to `BoundConcludabeResolver` so that when waiting for materialisations is finished messages are replayed in the expected order exactly.
- Upgrade the tracer to handle parallel requests
- Request traces are disabled when `traceInference` is off, for better performance.