Replace properties with YAML configuration (#6473)
## What is the goal of this PR?

In order to facilitate our long term goals of having a single configuration system for the entire database, including logging, we convert the preexisting `typedb.properties` and `logback.xml` into a Yaml single configuration file. This will also allow us to replicate configurations across an entire cluster in the future. In the process we also fix the bug where we were unable to configure log directories to be outside of the server distribution (#6370, #6372).

The top-level API is almost the same as before, please use the help command to get an overview:
```
./typedb server --help
```

Note: all YAML configurations can be overriden or extended with command line args, with `.` for nesting:
```
./typedb server --log.output.file.directory=/var/log/typedb
```
will override the persisted log directory set in the configuration.

The medium term goal is to eliminate the user-configurable `typedb.yml` file that comes with the server and ask the user to update configurations directly via console (https://github.com/vaticle/typedb/issues/6461).

## What are the changes implemented in this PR?

* Introduce the idea of YAML-shaped configurations
* to set a configuration, either override it from the command line with `--<option>=<value>`, or provide a custom, complete YAML configuration file from an external place and set it via `--configFile=<path-to-config-file>`, or update the configuration file that ships with the distribution
* Logging is now configured at bootup by the server, accoring to the configuration
* we replace PicoCLI and implement our own parsing to use a YAML configuration file and allow named/dynamic keys to be provided
* implement our log configuration specification, along with Help menu printing, etc.
* Update the entry point of the `TypeDBServer` to consume a `Configuration`, and remove the old class `RunOptions`