Utilise IID type information in type inference (#6504)
## What is the goal of this PR?

We can use some of the information in IIDs by extracting the contained type IID and using that for type resolution, rather than ignoring IIDs completely. We had previous removed the code path within type inference that examined the IIDs within the query, following the logic that a static type checker should not validate the data specified within the query, only the structure & typing. This changed is outlined in https://github.com/vaticle/typedb/issues/6477.

## What are the changes implemented in this PR?

* extract type information from IIDs and include them in type inference
* add test to the suite of integration tests for type inference