Guarantee 'match' query answer uniqueness (#6513)
## What is the goal of this PR?

TypeDB guarantees uniqueness of `match` query answers in all cases. Previously, there were cases where violations of answer uniqueness occurred, such as when the server made particular query optimisations. 

## What are the changes implemented in this PR?

* use a `distinct()` set to guarantee concept map uniqueness
  * where the distinct set is applied depends on whether the start variable is  part of the variables returned to the user
  * if the start variable is not returned, we have to use a more expensive larger distinct set to guarantee uniqueness - ther is room for optimisation in the future.

## Example of bug:

If this these two answers exist, trying to return just `$x`:
```
{$_0: 0x10, $x: 0x2}
{$_0: 0x10, $x: 0x2}
```

If the start variable is `$_0`, then the old code would only use a distinct set for the remainder of the answer, so we end up with two answers of `{$x: 0x2}`.
If the start variable was `$x`, we would end up with just one answer of `{$x: 0x2}`.