Significantly faster rule conclusion insertability validation (#6521)
## What is the goal of this PR?

Rule validation that checks each permutation of types from the `when` of the rule is insertable in the `then` of the rule is made much more performant. This allows very widely applicable rules to be validated and committed in a reasonable amount of time. Previously, when a rule could insert many permutations of types (>100k) the validation could appear to stall the server.

## What are the changes implemented in this PR?

* we pass a filter into type inference, so it returns fewer type permutations
  * when the query planning gets smarter, we can make this even faster by planning the variables we want to return earlier in the plan. This will allow us to switch from overgenerate and filter, to generate fewer permutations in the first place
* we use a set `contains()` instead of `O(n)` check to see whether a permutation is valid. This hugely improves performance when there are many valid permutations, eg. > 10k that have to be checked for each of >10k when type permutations