Fix role player statistics used in query optimisation (#6558)
## What is the goal of this PR?

We fix a minor bug with the calculation of role player statistics in query optimisation.

## What are the changes implemented in this PR?

When calculating the cost of traversing a RolePlayer edge from the relation to the player, we want to divide the number of instances of the role in question by the number of instances of all relations that relate the role. The latter was originally counting the number of instances of the scope of the role (such as "parenthood" in "parenthood:parent"). That undercounts the number of relations by not considering the relations which are instances of subtypes of the relation. The worst case occurs when the original relation type has no instances associated with it.