Rewrite query engine to leverage data sortedness (#6546)
## What is the goal of this PR?

We implement a new, non-recursive traversal algorithm that takes advantage of TypeDB's natively sorted data layout to (potentially dramatically) improve query performance. This is especially useful in queries with multiple meeting points which can make use of intersections over the sorted data streams.

The new architecture is also ready to leverage multiple starting points in queries should the query planner provide it.

## What are the changes implemented in this PR?

We have fundamentally changed the the traversal engine in several ways:
1. We utilise a different type of "join" for traversing vertices, which uses sortedness. As part of this we change from an edge-by-edge traversal to a vertex-by-vertex traversal.
2. We use a non-recursive backtracking algorithm instead of a recursive backtracking algorithm, which allows us to short circuit backtracking in tree-shaped queries more easily and without patchy data structures
3. We cannot filter out duplicate roles in the traversal during edge iteration, because we no longer traverse edge-by-edge, but instead filter them out when we visit each vertex as a separate step.

### Algorithmic changes

**Intersection-based vertex exploration**

We can consider this algorithm as roughly equivalent to implementing a sophisticated `merge-join` (https://en.wikipedia.org/wiki/Sort-merge_join).

By using sorted `Forwardable` data iterators, which can seek through the data to a particular point, we can treat all query graph meeting points as intersections over sorted iterators. This is combined with the previous backtracking algorithm, to explore the next query vertex using intersections over iterators from the previously explored and held fixed data vertices.

This implementation flips the query graph exploration from being a backtracking algorithm that explores edge-by-edge to a model that goes vertex-by-vertex. Any number of edges that lead into a new vertex being explored are intersected at the same time. 

Note: In the future, we may choose not to use only intersections when one of the iterators is small, since the underlying RocksDB `seek()` is slower than just doing a `get()`.

The logic of the actual traversal, and checking the validity of intersections, recording where to backtrack to and revisit, is performed at the top level of the `GraphIterator`. The vertex iterators are held by `GraphIterator.VertexTraverser` - these objects manage iterator cleanup and cache the last known intersection for use when re-trying.

**Non-recursive query graph exploration**

We also convert the previous recursive algorithm into a nonrecursive one. It operates in a "traverse" or "backtrack" mode,and in forward exploration always traverses the next lowest of the ordered vertices in a "traverse" set (vertices that must be visited next to find an answer). Or, at a dead end, it consumes the last vertex in the "revisit" set (vertices that have caused a dead end and need to be retried).

Utilising these sets naturally allows the algorithm to short circuit through the plan in tree-shaped queries where part of the exploration are isolated from each other. For example:
```
0: $a [name]
1: $x [person]
  $a -has-> $x
2: $y [dog]
  $a -has-> $y
```
If we fail when exploring `$y`, we can skip back to `$a` without going to `$x` in between. Similarly, in a tree multiple layers deep, we can fast-forward to the children of changed vertices without exploring all the interleaved ones. Abstract example:
```
     0
  /      \
1        2 
|         |
3       4
```
In this case, if `4` fails to find an answer, we can immediately generate the next `2` without changing `3` first. This does moderately change the output of the query plan, however we know that in general DFS in tree-shaped queries is cheaper than a layered BFS approach, which opens more iterators overall. Additionally, this approach is the fastest way to answer a query with no answers (in this example, if there are no 2's with a 4 at all, following the plan to find other 3's has no effect at all!)
This makes us slightly more immune to query planning irregularities at no cost.

Compare this to the recursive algorithm, which implemented the backtracking short circuiting by maintaining complex state from a dead end, in order to skip through unrelated vertices to get to previous vertices that could have caused the dead end. In the new model, we can do both forward and backward skipping by nature of the implementation without specialised state.

**Traversing roles**

We no longer check and filter duplicate roles in a relation's scope at generate time (ie. within the edge iterators).
Instead, we go over the edges regardless of the role, record the role, and then validate no duplicates exist separately at each vertex. To see why this had to be changed, please read: https://github.com/vaticle/typedb/pull/6546/files#r894024814

Note: this is the mechanism that allows us to query `($x, $y)` and get all permutations of `$x` and `$y` without them being the same role players. 

Additionally, because we no longer do edge-by-edge forward and back-tracking, we cannot answer queries like `$r ($x, $x)` by doing an intersection over two identical iterators on `$r` - we would intersect correctly, but would also record identical roles in the process. To solve this, we now de-optimise one of the duplicate role players queries by traversing one `$x --RolePlayer--> $r` optimised edge, and visit one role vertex as well: `$x --Playing--> $_role:$r` and `$r --Relating--> $_role:$r`.  This means we can backtrack to a vertex and get the required answer permutations.

### Smaller changes

* Separate `loop` edges in Traversal objects (including in all of `Traversal` subtypes: `Structure`, `Planner`, and `Procedure`) from `ins` and `outs`.
  * in most places, we were filtering out/specially handling loop edges already. Splitting them into a separate structure makes the special cases more obvious and introduces fewer `if` checks throughout the traversal package
  * NOTE: remaining code smell -`Planner` edges currently MUST be `Directional`, which `loop` edges don't have to be.
* `GraphProcedure` has been changed to be constructed via the `Builder` inner class. Previously, we just used a static factory method to create the `GraphProcedure` from a `GraphPlanner`. 
  * This allows the `GraphProcedure` just to accept an ordered array of `ProcedureVertex`, and the builder takes care of registering and ordering vertices before the `GraphProcedure` is created. This will be simplified after the query planner orders vertices natively.
* We used to restrict, by typing, that only `ThingVertex` was comparable to `ThingVertex`, and `TypeVertex` was comparable to `TypeVertex` for sortedness. However, we now just allow any `Vertex` to be comparable to any `Vertex`, simplifying some generics
  * this was required because the traversal engine only operates over intersections of generic `Vertex` iterators
* Cache hash codes in `Procedure` objects - these hash codes end up incurring a real cost because they are used so often
* Fix various bugs in laziness and evaluation of `Forwardable` iterators