Fix intermittent incomplete type matches (#6590)
Backward ISA bugfix## What is the goal of this PR?

Fix a bug that caused an `isa` query where the type is a variable to occasionally match fewer types than expected.

## What are the changes implemented in this PR?

A typo in ProcedureEdge caused the traversal to only consider direct types of a thing when traversing the `isa` edge when coming from the type. It appears that this is a rarely chosen direction of an `isa` edge in the planning (and, therefore, in the behaviour test suite), so an explicit test is added.