Improve negative-cycle detection during rule-validation (#6594)
## What is the goal of this PR?
The part of rule validation which prevents negative cycles (validateRulesThroughNegativeCycles) currently:
1.  Hangs on rule-commit in certain situations because of an infinite loop (#6500)
2. Falsely flags rules without negative cycles (#6599)

This PR fixes both.

## What are the changes implemented in this PR?
When traversing the dependency tree, the set of visited nodes was tracked using the valueSet of a back-pointer map. Duplicate keys caused an overwrite of certain nodes, allowing infinite looping. This is fixed by using the keyset.

Any cycle involving a rule containing negated concludables was flagged. This was narrowed down to only flag rules where the cycle starts at a negated concludable.

- fixes #6500 
- fixes #6599