Correctly add RocksDB JAR to the distribution (#6598)
## What is the goal of this PR?

We updated our distribution targets to ensure the RocksDB JAR is correctly added.

## What are the changes implemented in this PR?

The RocksDB JAR is added to the output distribution via its own `java_deps` target, named `//server:server-deps-prod`. `java_deps` used to package both the `data_runfiles` and `files` of the `java_library` and `java_binary` targets passed to it. This was largely superfluous, as when called on a `java_binary`, `data_runfiles` would contain all necessary files. But `:server-deps-prod` was incidentally working because of it.

To fix it, we've introduced a `java_binary` target, `:server-bin-deps-prod`, in line with our other uses of `java_deps` in this repo (e.g: `server-bin-deps-mac`.