Fix erroneous session timeouts (#6595)
## What is the goal of this PR?

We fix a bug which allowed session idle timeouts to trigger unexpectedly.

## What are the changes implemented in this PR?

If `mayStartIdleTimeout()` was called at the same time in two different threads, as long as the session had no open transactions, two timeouts were scheduled. The object only holds on to one reference though, so if a new transaction is opened, only one future is cancelled. If, when the timer runs out, no transactions happen to be open, the other future—which could never be cancelled—would close the session. From the outside that looks like occasionally the timeout triggers immediately (within milliseconds) after the last transaction is closed.

We now refuse to schedule a timeout if one is already scheduled.