Add retries to SoundnessVerifier for reasoner soundness tests (#6565)
## What is the goal of this PR?
Behaviour tests for the soundness of the reasoner fail when they involve cyclic explanations. This was due to the incorrect mapping of inferred concepts between the reasoner's transaction and the oracle's transaction.

## What are the changes implemented in this PR?

The SoundnessVerifier can only correctly map an inferred concept 'C' when it processes an explanation where
1. All the concepts in the condition of the rule are mapped
2. C occurs in the conclusion of the rule3. 

This repeatedly processes the explanation set, processing only explanations which satisfy the above condition. 
Each iteration verifies some explanations, enabling others to be verified in the following iteration, until all explanations have been processed. If progress is not made, an exception is thrown.